validacoes funcionando

+++ src/controllers/caixa-controller.js
@@ -1,5 +1,7 @@
 const Pagamento = require('../model/caixa-model.js')
 const Caixa_DAO = require('../DAO/caixa-DAO')
+const ConfirmaID = require('../model/confirma-id')
+const ConfirmaPedido = require('../model/confirma-pedido')
 
 const caixa = (app, bdCaixa) => {
     const caixa_Dao = new Caixa_DAO(bdCaixa);
@@ -25,9 +27,11 @@ const caixa = (app, bdCaixa) => {
 
     // })
 
+
     app.get('/caixa/:pedido', async (req, res) => {
         try {
             const pedido = req.params.pedido
+            console.log("pedido",pedido)
             const respCaixaPedido = await caixa_Dao.select_caixa_pedido(pedido);
             res.status(200).json(respCaixaPedido)
         } catch (error) {
@@ -39,7 +43,7 @@ const caixa = (app, bdCaixa) => {
     app.post('/caixa', async (req, res) => {
         const body = req.body
         try {
-            
+
             const novoPagamento = new Pagamento(body.PEDIDO, body.VALOR)
             const respNovoPedido = await caixa_Dao.insert_caixa_pedido(novoPagamento);
             res.status(200).json(respNovoPedido)
@@ -50,13 +54,28 @@ const caixa = (app, bdCaixa) => {
     })
 
     app.put('/caixa/:pedido', async (req, res) => {
-        const pedido = req.params.pedido 
+        const pedido = req.params.pedido
         const body = req.body
         try {
-            
-            const novoPagamento = new Pagamento(body.PEDIDO, body.VALOR)
-            const respNovoPedido = await caixa_Dao.update_caixa_pedido(pedido, novoPagamento);
-            res.status(200).json(respNovoPedido)
+            const pedido = req.params.pedido
+            const respCaixaPedido = await caixa_Dao.select_caixa_pedido(pedido);
+            const confPedido = new ConfirmaID(respCaixaPedido.pedidos.length)
+            console.log(respCaixaPedido.pedidos.length)
+            if (confPedido.id === -1) {
+                res.send("Item não encontrado")
+            } else {
+                const novoPagamento = new Pagamento(body.PEDIDO, body.VALOR)
+                if (novoPagamento.pedido === 0) {
+                    res.send("Campo pedido é obrigatorio")
+                } else if (novoPagamento.valor === -1) {
+                    res.send("Campo Valor é obrigatorio")
+                } else {
+                    console.log(novoPagamento)
+                    const respNovoPedido = await caixa_Dao.update_caixa_pedido(pedido, novoPagamento);
+                    res.status(200).json({ "pedido": novoPagamento, "message": "alterado conm sucesso!!" })
+                }
+            }
+
         } catch (error) {
             res.status(404).json({ error })
         }
@@ -64,12 +83,18 @@ const caixa = (app, bdCaixa) => {
     })
 
     app.delete('/caixa/:id', async (req, res) => {
-        const id = req.params.id 
-        console.log(id)
+        const id = req.params.id
         //const body = req.body
         try {
-            const removePedido = await caixa_Dao.remove_caixa_pedido(id);
-            res.status(200).json(removePedido)
+            const idRetorno = await caixa_Dao.select_caixa_id(id);
+            console.log(idRetorno.pedidos.length)
+            const confId = new ConfirmaID(idRetorno.pedidos.length)
+            if (confId.id === -1) {
+                res.send("Item não encontrado")
+            } else {
+                const removePedido = await caixa_Dao.remove_caixa_pedido(id);
+                res.status(200).json(removePedido)
+            }
         } catch (error) {
             res.status(404).json({ error })
         }
